requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
streamlit-extras>=0.3.0
plotly>=5.15.0
pillow>=10.0.0
openai>=1.0.0
numpy>=1.24.0
firebase-admin>=6.2.0
pyrebase4>=4.6.0

